+++ src/test/java/com/java/CountNumberOfDigits.java
@@ -1,0 +1,63 @@
+package com.java;
+
+public class CountNumberOfDigits {
+
+	public static void main(String[] args) {
+		
+		//Count number of Digits in a number
+		int num=123456;
+		int count=0;
+		
+		while(num>0) {
+			num=num/10;
+			count++;
+		}
+		System.out.println(count);
+	}
+}
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+		
+//		
+//		int num=123456;
+//		int count=0;
+//		
+//		while(num>0) {
+//			num=num/10;
+//			count++;
+//		}
+//		System.out.println("Number of digits in a number is: " + count);
+
+
